add: 유저 목록 검색 API Router & NetworkManger로 변경
